Use Libsass to compile SCSS to CSS
Since the beginning of gulp workflow, Suit-baze uses gulp-ruby-sass
(which using Ruby) to compile SCSS to CSS. It was great until the
compilation time hits 7s.

Libsass in other hand has good performance in exchange of features
available. Seeing the complexity of current projects, we, front-end
devs of Suitmedia can live with that.